requirements.txt
requests
python-telegram-bot[job-queue]
pyalex
python-dotenv
pyyaml
openpyxl
sqlalchemy
gigachat
beautifulsoup4
playwright
PyMuPDF
readability-lxml
apscheduler
arxiv
pylatexenc